pyproject.toml
[project]
name = "eat-it"
version = "0.1.0"
description = "Recipe Storage and Smart Shopping List"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [{ name = "Luke McGuire", email = "luke.mcguire@gmail.com" }]
dependencies = [
    "sqlmodel>=0.0.16",
    "alembic>=1.13",
    "fastapi[standard]>=0.115",
    "uvicorn>=0.30",
    "aiosqlite>=0.20",
    "sqlite-vec>=0.1.0",
    "sentence-transformers>=3.0",
    "pydantic-settings>=2.0",
    "recipe-scrapers>=15.11.0",
    "httpx>=0.28.1",
    "ingredient-parser-nlp>=2.5.0",
]

[project.entry-points."eat_it.importers"]
# Built-in importers will be added in Phase 2

[project.urls]
Homepage = "https://github.com/lukemcguire/eat-it"
Repository = "https://github.com/lukemcguire/eat-it"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ─── Hatch Build Configuration ──────────────────────────
[tool.hatch.build.targets.wheel]
packages = ["src/eat_it"]

# ─── Ruff ─────────────────────────────────────────────
[tool.ruff]
line-length = 120
fix = true
extend-exclude = ["**/.venv"]

[tool.ruff.lint]
select = [
    "FAST", # FastAPI - FastAPI specific rules
    "YTT",  # flake8-2020 - checks for Python 2020 best practices
    "ANN",  # flake8-annotations - enforces type annotation style
    "S",    # flake8-bandit - security issues
    "BLE",  # flake8-blind-except – flags bare excepts
    "FBT",  # flake8-boolean-trap – potential pitfalls with booleans
    "B",    # flake8-bugbear - common bug patterns
    "A",    # flake8-builtins - misuse of python builtin names
    "C4",   # flake8-comprehensions - best practices in comprehensions
    "DTZ",  # flake8-datetimez – requires timezone-aware datetime objects
    "T10",  # flake8-debugger – debugger statements (e.g. pdb)
    "SIM",  # flake8-simplify
    "EM",   # flake8-errmsg – error message style
    "EXE",  # flake8-executable – executable file checks
    "FA",   # flake8-future-annotations – future import for annotations
    "ISC",  # flake8-implicit-str-concat – warns on implicit string concatenation
    "ICN",  # flake8-import-conventions – enforces conventional import aliases
    "LOG",  # flake8-logging – proper logging usage
    "G",    # flake8-logging-format – logging format string issues
    "INP",  # flake8-no-pep420 – warns against non-PEP420 namespace usage
    "PIE",  # flake8-pie – Python improvement suggestions
    "PYI",  # flake8-pyi – checks for type stub (.pyi) consistency
    "PT",   # flake8-pytest-style – pytest best practices
    "Q",    # flake8-quotes – enforces quote style consistency
    "RSE",  # flake8-raise – proper raise statement usage
    "RET",  # flake8-return – return statement issues
    "SLF",  # flake8-self – flags instance methods that don't use self
    "SLOT", # flake8-slots – suggests use of __slots__ where appropriate
    "TID",  # flake8-tidy-imports – enforces specific import styles (e.g., no relative imports)
    "TC",   # flake8-type-checking – proper import of typing in type checks
    "INT",  # flake8-gettext – checks for proper internationalization usage
    "ARG",  # flake8-unused-arguments – flags unused function arguments
    "PTH",  # flake8-use-pathlib – encourages pathlib over os.path
    "TD",   # flake8-todos – flags TODO comments
    "FIX",  # flake8-fixme – flags FIXME comments
    "FLY",  # Flynt - f-string conversion suggestions
    "I",    # isort - import ordering checks
    "C90",  # mccabe - complexity metric for functions
    "NPY",  # NumPy-specific rules - ensures NumPy coding standards
    "PD",   # pandas-vet – checks pandas-specific code practices
    "N",    # PEP8 Naming – naming conventions
    "PERF", # Perflint – performance-related checks
    "E",    # pycodestyle errors (style issues)
    "W",    # pycodestyle warnings (style issues)
    "D",    # pydocstyle - docstring formatting
    "F",    # pyflakes - detects syntax errors and basic mistakes
    "PGH",  # pygrep-hooks - custom grep hooks for linting
    "UP",   # pyupgrade - upgrades syntax to newer Python versions
    "RUF",  # Ruff-specific rules - additional ruff checks
    "TRY",  # Tryceratops – try/except usage suggestions
]

ignore = [
    "E501", # line-too-long
    "S311", # suspicious-non-cryptographic-random-usage
    "D107", # undocumented-public-init
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "D",      # pydocstyle
    "S101",   # assert-used
    "ANN",    # annotations
    "BLE",    # blind-except
    "TRY",    # tryceratops
    "INP001", # implicit-namespace-package
    "SLF001", # private-member-access
    "S603",   # subprocess-without-shell-equals-true
    "N802",   # invalid-function-name (test names like test_NEW_01)
    "ARG001", # unused-function-argument (mock signatures)
    "ARG002", # unused-method-argument (pytest fixtures)
    "ARG005", # unused-lambda-argument (MagicMock protocol)
    "DTZ001", # call-datetime-without-tzinfo (test fixtures)
    "DTZ005", # call-datetime-now-without-tz (test fixtures)
    "PT011",  # pytest-raises-too-broad
    "RUF059", # unpacked-variable-unused
    "SIM117", # multiple-with-statements (test fixtures)
    "SIM118", # in-dict-keys
]

# ─── Type Checking ────────────────────────────────────
[tool.ty.environment]
python = "./.venv"
python-version = "3.13"

# ─── Pytest ───────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--import-mode=importlib --timeout=60"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

# ─── Coverage ─────────────────────────────────────────
[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["."]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings>=1.0.3",
    "mkdocstrings-python>=2.0.2",
    "prek>=0.3.2",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.12",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.0",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]
